refactor: unify DTO ID types to Long for consistency with DB and domain

+++ src/main/java/com/kbstar/littlestar/user/dto/GatchaPokemonRequest.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class GatchaPokemonRequest {
 	@NotNull
-	private Integer newPokemonId;
+	private Long newPokemonId;
 }
